move tests to integration
